graphrag/streamlit/requirements.txt
streamlit==1.36.0
streamlit-feedback==0.1.3
langchain==0.2.1
langchainhub==0.1.17
langchain-core==0.2.3
langchain-community==0.2.1
neo4j==5.20.0
ollama==0.2.0
python-dotenv==1.0.1
